[Debug] Correction of a big bug and a potential one.
The big bug was that the ifstream and ofstream used when read/writing
binary format were never set to binary mode. This is not such a problem
on Unix, but is create a weird bug on MS Windows. On this platform, if
the binary mode is set, the file is re-opened with the proper mode. This
is the best fix I found.

Another potential threat was that the reading of a file in binary mode
could lead to reading more bytes than the storage of the allocated
vector. Now the remaing size to fill is accounted for.