build: Rename 'OPENMP_FLAG'; remove 'OPENMP_LIBS'.
'OPENMP_FLAG' is now 'OPENMP_CFLAGS'.
'OPENMP_LIBS' has been removed in favor of 'OPENMP_LINKFLAGS'.

The rationale is that the documented way to use OpenMP is via
compiler *and* linker flags, usually '-fopenmp'.  Explicitly linking
against the compiler's OpenMP library may or may not work, and the
library name is an implementation detail that shouldn't be relied on.